Move state logic to useDeleteUser
