+++ packages/react/src/components/AccountSettings/DeleteUser/DeleteUser.tsx
@@ -1,12 +1,12 @@
 import React from 'react';
 
-import { deleteUser, translate, getLogger } from '@aws-amplify/ui';
-import { useAuth } from '@aws-amplify/ui-react-core';
+import { translate, getLogger } from '@aws-amplify/ui';
+import { useDeleteUser } from '@aws-amplify/ui-react-core';
 
 import { Flex } from '../../../primitives';
 import { ComponentClassName } from '../types';
 import DEFAULTS from './defaults';
-import { DeleteUserProps, DeleteUserState } from './types';
+import { DeleteUserProps } from './types';
 
 const logger = getLogger('Auth');
 
@@ -16,60 +16,28 @@ function DeleteUser({
   onError,
   handleDelete,
 }: DeleteUserProps): JSX.Element | null {
-  const [state, setState] = React.useState<DeleteUserState>('IDLE');
-  const [errorMessage, setErrorMessage] = React.useState<string>(null);
+  const {
+    state,
+    errorMessage,
+    user,
+    isLoading,
+    startConfirmation,
+    handleCancel,
+    handleConfirmDelete,
+  } = useDeleteUser({
+    onSuccess,
+    onError,
+    handleDelete,
+  });
 
   // translations
   const deleteAccountText = translate('Delete Account');
-
-  const { user, isLoading } = useAuth();
 
   // subcomponents
   const { ErrorMessage, DeleteButton, WarningView } = React.useMemo(
     () => ({ ...DEFAULTS, ...(components ?? {}) }),
     [components]
   );
-
-  const startConfirmation = (event: React.MouseEvent<HTMLButtonElement>) => {
-    event.preventDefault();
-    setState('CONFIRMATION');
-  };
-
-  const runDeleteUser = React.useCallback(async () => {
-    setState('DELETING');
-    if (errorMessage) {
-      setErrorMessage(null);
-    }
-
-    try {
-      if (handleDelete) {
-        /*
-         * run custom delete handler, if provided. We pass `user` so that
-         * developer can do whichever cleanup with the user object they wish.
-         */
-        await handleDelete(user);
-      } else {
-        // else, run default deleteUser function.
-        await deleteUser();
-      }
-      setState('DONE');
-      onSuccess?.();
-    } catch (e) {
-      const error = e as Error;
-      setState('ERROR');
-      setErrorMessage(error.message);
-      onError?.(error);
-    }
-  }, [errorMessage, handleDelete, onError, onSuccess, user]);
-
-  // called when end user cancels account deletion confirmation
-  const handleCancel = React.useCallback(() => {
-    setState('IDLE');
-  }, []);
-
-  const handleConfirmDelete = React.useCallback(() => {
-    runDeleteUser();
-  }, [runDeleteUser]);
 
   // Return null if Auth.getCurrentAuthenticatedUser is still in progress
   if (isLoading) {
